Add an installDebFile method to QApt::Backend and the QApt Worker.
This method starts the installation of the given .deb file, complete with a DebInstallStarted event signal on start, DebInstallFinished when finished, and a debInstallMessage() signal in between for dpkg output.
A new error, WrongArchError, has been added for when the .deb file cannot be installed due to incompatible CPU architectures for binary packages.