More fixes for MultiArch. The last round of fixes were made with the assumption that all installable foreign-arch packages would be marked with a Multi-Arch field in their control files. This was a bad assumption. Non-native foreign-arch packages need no changes made to them as long as their dependencies are multi-arch.
As such, we cannot only add foreign-arch packages marked as MultiArch to the cache, or else
we lose foreign-arch binary packages such as Skype. (Which is what people actually want!)

Unfortunately, this means that there will be "duplicate" entries for most packages in the
cache, but these can be differentiated by placing (arch) after the names in your display
model.

Package::isMultiArchEnabled() has been deprecated, and a new Package::isForeignArch()
function has been added. It does the same thing, but includes all foreign-arch packages
that the APT install supports.