Fix build with newer versions of Xapian.
More recent versions of Xapian have a class with a member variable called "slots".
This conflicts with a macro that Qt defines also called slots. We are already using
the safer Q_SLOTS macro in LibQApt, but we still need to #undef slots, since Qt
declares the macro whether we use it or not.

BUG:280024
FIXED-IN:1.2.3