directly include declaration of MD5Summation
While working on apt 0.9 I cleaned up the header includes. I thought
I had guarded these changes with APT_8_CLEANER_HEADERS, but it seems
I forgot one, which makes qapt FTBFS as it depends on such a transitive
include. As it is a good idea to avoid these transitive includes anyway
lets do the include in qapt instead of readding it to apt.