Add a stub for the client-side Transaction class. It is an implicitly-shared class that represents a transaction object on D-Bus from the QApt Worker. Clients cannot directly effect the data members of the Transaction class, so data is never copied.
In the future, QApt::Backend will have a transaction list. It will remove its own transaction instance when the transaction finishes, and when clients do the same with their own copy(ies), the data backing the instance will have a refcount of zero and be deleted.
The user won't have to worry about memory leaks, and the class makes dealing with transactions very stable in terms of potential segfaults.

If you couldn't tell, I'm pretty pleased with myself. QSharedPointer is kind of awesome.