In some cases QApt::Backend will check for a state flag of a Package that will not change over the life of the Package. The fact that some flags will not change is capitalized upon by the Package class, which memoizes a set of immutable state flags the first time Package::state() is called. Unfortunately, until this commit the only way to access these state flags was to call Package::state(), which returns the composite of the immutable state flags, as well as calculations to return the current values of all the mutable flags.
As a small optimization, I added a (private) staticState() method to Package so that Backend can check the value of a known-immutable state without invoking the calculation of the other mutable flags.
This provides a nice speed bump for certain operations, including Backend::upgradeablePackages() and Backend::markPackages(), which check a package's state for static flags.