DebFilePrivate defaults to invalid & prevent unhandled exceptions on open
debDebFile::MemControlExtract has potential to run into unhandled out of
range exceptions when encounting a path/file that is not valid (by calling
substr on a nullstr it appears).
to prevent this we now use the only way one works with exceptions ... by
catching (...) and considering the debfile invalid whenever an exception
occurs.

also the debfile now defaults to !valid after construction (was valid) to
require explicit marking as valid rather than explicit marking as invalid,
as unless we actually proof it is valid we must consider it invalid.