fix return value on success
previously when reject was called with m_done=true it would call accept
which does nothing on qdialog. instead call qapp::quit which will exit
with 0. exit(0) is necessary for gstreamer to know that installation was
executed succesfully, only then will the plugin cache be reloaded and
the media gets retried (i.e. this fix enables phonon gstreamer applications
to request codec install and on successful exit they reload the cache
and play the media as if nothing happened)