adjust gst version matching to possibly enable gst0.10 support
with x.y as required version for the plugin install, make sure that
x is equal between what is requested and what is offered by the package.

e.g. gstreamer requests 0.10 plugin compatibility:
x=0, y=10;
for a package to be considered a viable candidate x must be exactly 0 and y
must be >= 10.

This prevents requests for 0.10 to be met by 1.x packages, while still
allowing any 1.x package to meet any 1.x requst. 0.10 packages continue to
not meet a 1.x request though.