Prevent crashes on source entries that contain a line like "##"
Simply ensure that while we are stripping # the string retains the
QString::at(i) requirement of i being a valid index (which requires the
size() to be greater than 0 for us).

Random note: while size>0 is equal to !isEmpty, the ::at() documentation
specifically ties the validity of index to size, so for clarity's sake I am
using size>0 rather than !isEmpty.

FIXED-IN: 2.2.1
BUG: 333675