fix package remove resolver
relevant discussion thread that lead to the breakage:
https://lists.ubuntu.com/archives/kubuntu-devel/2013-December/007623.html

The actually problem is a pile of madness. Originally the remove/purge
functions were derived from synatic which set up the resolver, called
the now deprecated ::InstallProtect, resolved and then marked the cache.
Removing InstallProtect however lead to incorrect resolution and ultimately
resulted in broken dependencies when removing a package. This is due
to what InstallProtect does. While the docs say it is no op it is not
really. It explicitly marks the packages marked for install/delete in the
resolver in the actually depCache, since the removal of InstallProtect the
depCache marking only happened *after* the resolver resolved. This
effectively meant that the cache marking worked without resolution
ultimately resulting in a broken cache.

The solution to this is obviously that the cache needs to be marked before
we let the resolver poke it with a stick.

(the original commit removing the installprotect calls also removes one in
 backend.cpp, but that one truly was noop as far as depCache is concerned)