Add Package::isInUpdatePhase to allow ubuntu-style update phasing
Update phasing is an additional feature provided by ubuntu/launchpad that
sets an additional control field called Phased-Update-Percentage. This
field defines a percentage of systems to which an update should be
delivered. This is to minimize exposure from update regressions by
gradually increasing the amount of systems to get this update.

To ensure even distribution of machines in any given percentage set of
phasing a dicrete randomizer with stable seed is used. A stable seed
based on packagesource-version-uniquemachineid is thrown at a simple
randomizer to get a percentage at which a given machine would fall into
the phasing, once the phasing defined by the remote repository exceeds the
machine specific phasing the query function returns true and a qapt client
*may* choose to honor this or not.

Whether or not phasing is going to be respected is a purely client specific
decision/setting. I.e. this also requires changes to muon-updater to
get phasing landed.

REVIEW: 121714