do not delay run response
setDelayedReply means we would have to issue a reply ourselves which from
what I can see never happens. So after a certain timeout (30 or so seconds)
dbus will automatically issue a NoReply error to the client. While this is
a truthy error in that there really was no reply in the time frame its
meaning client side is unclear as NoReply can then mean actually no reply
or everything was fine but no reply was issued. Latter appears to be the
success case because we never issue a reply, as already mentioned.
This in turn makes it impossible to tell client side when run simply took
too long or was successful.

To improve the situation simply do not mark the reply delayed. This makes
qtdbus send a default all-good reply to the client unless we issue an auth
error first.

Ideally the entire timeout scenario should be taken out of the picture by
employing an entirely async API design (i.e. run always replies immediately
but asyncronously switches state to either an error or running). The only
other option to prevent the noreply from being issued while waiting for
auth is to have the client timeout increased to indefinite which would seem
misguided as most of the operations have actual potential to result in
no-reply if something deadlocks or takes indeed way too long.

REVIEW: 123026