warn about state changes we cannot represent and exclude them from list
if a package changed state such that ToFoo nor Held are set we'd get
an empty standardized status we can't do anything with. so instead warn
and do not include this package in the list of changes. as far as changing
things goes having no ToFoo nor Held means whatever the state change
was, it is no-op anyway so reporting it as a change has no impact on
anything.

this presently happens with kubuntu ci where bluedevil cannot be installed
because one of its dependencies cannot be met.
when bluedevil however is already installed from the archive it is
originally upgradable and held but after marking it for dist-upgrade it
will simple be marked keep and not held. this is consistent with apt-get
and apt so apparently this is by design.

eitherway, in the event that a package changed state in an unexpected way
print a warning and its flags for possible inspection.