Add the needed compiler flags for gstreamer.
Recently in kubuntu wily the file gst/gstconfig.h was moved to
/usr/lib/$arch-linux-gnu/gstreamer-1.0/include/gst/gstconfig.h
so libqapt started to fail to build in wily because of this.

However pkg-config reports correctly the needed -I flags:
$ pkg-config --cflags-only-I gstreamer-1.0
-I/usr/include/gstreamer-1.0
-I/usr/lib/x86_64-linux-gnu/gstreamer-1.0/include
-I/usr/include/glib-2.0
-I/usr/lib/x86_64-linux-gnu/glib-2.0/include

The above flags, if I'm not mistaken are already stored in
${GSTREAMER_DEFINITIONS} by cmake/modules/FindGStreamer.cmake so the only
problem was that we were not using this variable to provide the compiler the
needed -I flags to the compiler.