Do not check status file for timestamp
Summary:
Update-apt-xapian-index is organised in plugins, each representing a
data source for the index. Each of them reports a timestamp of the
latest change, and the updated index takes the maximum of them as its
own timestamp.

We were checking the time of `Dir::State::status`
(`/var/lib/dpkg/status`), which is not used in any plugin. When it was
newer than all data sources, we would continuously see the index as out
of date.

BUG: 390476

Reviewers: sitter

Reviewed By: sitter

Differential Revision: https://phabricator.kde.org/D10715